Trying to fix with git ignore
