+++ Program.cs/Program.cs.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27130.2010
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Program.cs", "Program.cs\Program.cs.csproj", "{1A1EF94E-9EE3-44BF-AB36-BBB2E8111513}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1A1EF94E-9EE3-44BF-AB36-BBB2E8111513}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1A1EF94E-9EE3-44BF-AB36-BBB2E8111513}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1A1EF94E-9EE3-44BF-AB36-BBB2E8111513}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1A1EF94E-9EE3-44BF-AB36-BBB2E8111513}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {477C90F8-CF2D-4DAA-AFC5-F85DD32F663D}
+	EndGlobalSection
+EndGlobal
